Move helpers back to internal
