+++ internal/helpers/math.go
