+++ internal/helpers/struct.go
