+++ internal/helpers/struct_test.go
